# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MarketPay microfinance and USSD payment platform for informal market vendors in Sierra Leone.",
        "title": "MarketPay API",
        "contact": {
            "name": "MarketPay Engineering",
            "email": "engineering@marketpay.sl"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/reports/default-rate": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Default rate report",
                "responses": {}
            }
        },
        "/reports/disbursement-volume": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Disbursement volume by month",
                "responses": {}
            }
        },
        "/reports/officer-queue": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Loan officer pending review queue",
                "responses": {}
            }
        },
        "/reports/partner-summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Partner loans and commission summary",
                "responses": {}
            }
        },
        "/reports/portfolio": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Portfolio outstanding report",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/reports/repayment-rate": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Repayment rate report",
                "responses": {}
            }
        },
        "/reports/vendor-distribution": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Vendor distribution by market",
                "responses": {}
            }
        },
...
        "/vendors": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vendors"
                ],
                "summary": "List all vendors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vendors"
                ],
                "summary": "Register a new vendor",
                "parameters": [
                    {
                        "description": "Vendor data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_vendors_interfaces_http.createVendorRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_marketpay_backend_internal_vendors_domain_model.Vendor"
                        }
                    }
                }
            }
        },
        "/vendors/market-associations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "vendors"
                ],
                "summary": "List all market associations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_marketpay_backend_internal_vendors_domain_model.MarketAssociation"
                            }
                        }
                    }
                }
            }
        },
        "/vendors/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "vendors"
                ],
                "summary": "Get vendor by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_marketpay_backend_internal_vendors_domain_model.Vendor"
                        }
                    }
                }
            }
        },
        "/vendors/{id}/eligibility": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "vendors"
                ],
                "summary": "Check vendor loan eligibility",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/vendors/{id}/kyc/approve": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "vendors"
                ],
                "summary": "Approve vendor KYC",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_marketpay_backend_internal_vendors_domain_model.Vendor"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_marketpay_backend_internal_shared_domain_model.KYCStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "VERIFIED",
                "REJECTED"
            ],
            "x-enum-varnames": [
                "KYCPending",
                "KYCVerified",
                "KYCRejected"
            ]
        },
...
        "github_com_marketpay_backend_internal_vendors_domain_model.MarketAssociation": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_marketpay_backend_internal_vendors_domain_model.Vendor": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "business_name": {
                    "type": "string"
                },
                "business_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "credit_score": {
                    "type": "number"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "first_transaction_at": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kyc_status": {
                    "$ref": "#/definitions/github_com_marketpay_backend_internal_shared_domain_model.KYCStatus"
                },
                "kyc_verified_at": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "market_association": {
                    "$ref": "#/definitions/github_com_marketpay_backend_internal_vendors_domain_model.MarketAssociation"
                },
                "market_association_id": {
                    "type": "string"
                },
                "national_id_number": {
                    "type": "string"
                },
                "national_id_type": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_marketpay_backend_internal_vendors_domain_model.VendorStatus"
                },
                "transaction_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_marketpay_backend_internal_vendors_domain_model.VendorStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "ACTIVE",
                "SUSPENDED",
                "BLACKLISTED"
            ],
            "x-enum-varnames": [
                "VendorStatusPending",
                "VendorStatusActive",
                "VendorStatusSuspended",
                "VendorStatusBlacklist"
            ]
        },
...
        "internal_vendors_interfaces_http.createVendorRequest": {
            "type": "object",
            "required": [
                "date_of_birth",
                "first_name",
                "last_name",
                "market_association_id",
                "national_id_number",
                "national_id_type",
                "phone",
                "pin"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "business_name": {
                    "type": "string"
                },
                "business_type": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "market_association_id": {
                    "type": "string"
                },
                "national_id_number": {
                    "type": "string"
                },
                "national_id_type": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "pin": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}